chore(deps): import `JSX` type from `preact` instead of `preact/src/jsx`
the `package.json` of `preact` does not specify `src/jsx` within the `exports` property.
therefore, doing `import { JSXInternal } from 'preact/src/jsx'` does not
work in environments that use the `exports` field.

it works fine when `node_modules` are available in the project and
import path can be resolved. but for those using esm.sh or yarn with
pnp enabled the import path is not resolved, because for those package
managers `preact/src/jsx` is not available.

this commits changes the import to `import { JSX } from 'preact'`.
Internally, `preact` package re-exports the same `JSXInternal` type, so
it's the same thing, but now imported more correctly.