chore(calendar): add `type="button"` to all buttons (#837)
if user of schedule-x is rendering the calendar within a `<form>` tag,
then any of the calendar buttons become submit buttons by default.

A `<form><button>i submit</button></form>` is same as `<form><button type="submit">i submit</button></form>`.

this is a problem because the calendar buttons are not meant to submit,
for example the "Today" button is for changing the calendar view to the
current day. Currently if calendar is rendered within a form, clicking
the "Today" button will submit the form, which is unexpected for the
user.

Explicitly setting `type="button"` on all buttons within the calendar
components prevents this issue.

Although in some cases there is `event.preventDefault()` in the button
click handler, it's still worth setting `type="button"` to prevent form
submission (say if JS hasn't loaded yet, or if `event.preventDefault()`
is missed/removed.